fix(server): bring SQL fallback mode to parity and harden penpot import
- migrations add documentation_pages to SQLite/Postgres initial schema
- migration runner: comment-led tables + trigger handling; settings txn/JSON fixes
- search: non-Postgres LIKE fallback path
- penpot: reshape tokens import, new POST /api/penpot/components (dedupe by slug, status review)
- auth/status is mode-aware; client auth store follows PB or SQL mode

+++ server/repositories/component.repository.ts
@@ -190,7 +190,7 @@ class ComponentRepository {
     }> {
         const db = getDatabase()
         const result = await db.query<{ status: string; count: number }>(
-            `SELECT status, COUNT(*)::int as count 
+            `SELECT status, COUNT(*) as count 
        FROM components 
        WHERE deleted_at IS NULL 
        GROUP BY status`
